sounds in maprfid working on exe, resource path method moved to file utils from main
